Cleaning compilation errors and updating Catch2 dependency.

+++ examples/more_cube/render_system.cpp
@@ -107,7 +107,7 @@ void RenderSystem::update(float /*elapsedTime*/)
 	auto & pipeline = this->_process->getPipeline();
 	auto program = pipeline.getProgram();
 	auto lights = this->_scene.getLights();
-	program->bind(std::make_shared<ece::Uniform<int>>("numberOfLights", lights.size()), "numberOfLights");
+	program->bind(std::make_shared<ece::Uniform<int>>("numberOfLights", static_cast<int>(lights.size())), "numberOfLights");
 
 	int lightId = 0;
 	for (auto & light : lights) {
